requirements-dev.txt
numpy>=1.23
scipy>=1.16
pandas>=1.5
scikit-learn>=1.1,<1.6
joblib>=1.2
pytest>=7.4
flake8>=6.1
mypy>=1.6
jupyter>=1.0
datasets>=2.18
Pillow>=9.5
tqdm>=4.66

# Advanced training (optional - install for PyTorch-based diffusion models)
# pip install torch torchvision coremltools diffusers
torch>=2.0.0; platform_system != "Darwin" or platform_machine != "arm64"
torchvision>=0.15.0; platform_system != "Darwin" or platform_machine != "arm64"
coremltools>=7.0
diffusers>=0.25.0

# Apple Silicon specific versions
torch>=2.0.0; platform_system == "Darwin" and platform_machine == "arm64"
torchvision>=0.15.0; platform_system == "Darwin" and platform_machine == "arm64"
